Split sandbox/types.ts into per-subsystem type files
sandbox/types.ts co-changed with eight subsystems (local-environment,
sandbox-impl, sandbox-context, remote, tab-sync, persistence controller,
plus every consumer of the pyric/sandbox package export) making it a
shared-contract conflict surface. Regroup it under sandbox/types/, one
file per subsystem:

- auth-state.ts: AuthState (genuinely shared across all subsystems)
- errors.ts: SandboxErrorCode, DenialContext, SandboxErrorOptions, SandboxError
- persistence.ts: PersistableService, SandboxSnapshot
- events.ts: every SandboxEvent variant plus EventService/EventActor/AuthLens/EventProvenance
- service.ts: SandboxConfig, SandboxAdmin, Sandbox
- context.ts: SandboxContext

types/index.ts is a computed re-export barrel consumed only by the
package's public entry (sandbox/index.ts), which keeps its import path
unchanged for external consumers. Every other internal importer (6
files, including two using inline `import('../types.js').Foo` type
references in local-environment.ts that a static-import grep initially
missed) was updated to import directly from the concept file it needs,
per the no-internal-grab-bag-barrel rule.

No type shape changed. Verified: tsc build, pyric test suite (4560
pass), pyric-admin test suite (566 pass), lint:manifest all green.
